new pages
so I edited the frontcontent for the index page so it shows paintings that have been uploaded, and i figured out how to randomize and limit to 20!  then, i edited the painting.php? page so that it'll get the id of the painting you clicked on and display the info for that painting (and the artist of that painting haha which took me awhile)